neurotrade_ai_agent/requirements.txt
uagents>=0.21.0
aiohttp>=3.8.0
uagents-core>=0.3.0
aiohttp>=3.8.0
aiohttp-cors>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio
logging
requests>=2.31.0
websockets>=11.0.0

# The Graph specific dependencies
thegraph-python>=1.0.0
graphql-core>=3.2.0

# Additional trading dependencies
numpy>=1.24.0
pandas>=2.0.0
ccxt>=4.0.0

# Performance and monitoring
psutil>=5.9.0
prometheus-client>=0.16.0 